# swagger/v1.yml
swagger: '2.0'
info:
  title: flyte-api
  description: Self service platform automation API
  contact:
    name: FLYTE Team
    url: https://github.com/HotelsDotCom/flyte
  version: '1'
consumes:
  - application/json
  - application/x-yaml
produces:
  - application/json
  - application/x-yaml
paths:
...
  '/v1/packs/{packId}/actions/take':
    post:
      tags:
        - action
      summary: take action
      operationId: takeAction
      parameters:
        - $ref: '#/parameters/packId'
        - $ref: '#/parameters/commandName'
      responses:
        '200':
          description: action to be processed
          schema:
            $ref: '#/definitions/action'
...
definitions:
...
  link:
    type: object
    properties:
      href:
        type: string
      rel:
        type: string
    additionalProperties:
      type: string
...
  action:
    type: object
    properties:
      command:
        type: string
      input:
        type: object
      links:
        type: array
        items:
          $ref: '#/definitions/link'
...
parameters:
...
  commandName:
    name: commandName
    in: query
    description: command name
    required: false
    type: string

  # path parameters
  packId:
    name: packId
    in: path
    description: pack id
    required: true
    type: string
...